requirements.txt
gradio
Pillow
rdkit
scikit-learn